Add new tests for multi-condition and convert old tests.
