+++ test/test_market_makers.js
@@ -33,22 +33,21 @@ contract('MarketMaker', function(accounts) {
         
         const conditionId = await getParamFromTxEvent(
             await pmSystem.prepareCondition(oracleAddress, questionId, numOutcomes),
-            'conditionId')
+        'conditionId')
 
-            const investor = 0
-            const feeFactor = 0  // 0%
+        const investor = 0
+        const feeFactor = 0  // 0%
      
         // Create and fund lmsrMarketMaker
         const funding = toBN(1e17)
         await etherToken.deposit({ value: funding, from: accounts[investor] })
         await etherToken.approve(checksummedLMSRAddress, funding, { from: accounts[investor] })
-        assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), funding.toString())
+        assert.equal(await etherToken.balanceOf(accounts[investor]).then(v => v.toString()), funding.toString())
+        
         const lmsrMarketMaker = await getParamFromTxEvent(
-            await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, toBN(1e17),
-                { from: accounts[investor] }),
-                'lmsrMarketMaker', LMSRMarketMaker)
+            await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, toBN(1e17),{ from: accounts[investor] }), 'lmsrMarketMaker', LMSRMarketMaker)
         nonce++;
-        assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), '0')
+        assert.equal(await etherToken.balanceOf(accounts[investor]).then(v => v.toString()), '0')
         
         // User buys all outcomes
         const trader = 1
@@ -63,12 +62,13 @@ contract('MarketMaker', function(accounts) {
         await pmSystem.setApprovalForAll(lmsrMarketMaker.address, true, { from: accounts[trader] })
 
         // User sells tokens
-        const buyerBalance = await etherToken.balanceOf.call(accounts[trader])
+        const buyerBalance = await etherToken.balanceOf(accounts[trader])
+        
         let profit, outcomeTokenAmounts
         for(const i of _.range(loopCount.toNumber())) {
             // Calculate profit for selling tokens
             outcomeTokenAmounts = Array.from({length: numOutcomes}, (v, i) => i === outcome ? tokenCount.neg() : toBN(0))
-            profit = (await lmsrMarketMaker.calcNetCost.call(outcomeTokenAmounts)).neg()
+            profit = (await lmsrMarketMaker.calcNetCost(outcomeTokenAmounts)).neg()
             if(profit.eqn(0))
                 break
 
@@ -79,7 +79,7 @@ contract('MarketMaker', function(accounts) {
 
             netOutcomeTokensSold[outcome] -= tokenCountRaw
             const expected = lmsrMarginalPrice(funding, netOutcomeTokensSold, outcome)
-            const actual = new Decimal(await lmsrMarketMaker.calcMarginalPrice.call(toBN(outcome)).then(v => v.toString())).div(ONE)
+            const actual = new Decimal(await lmsrMarketMaker.calcMarginalPrice(toBN(outcome)).then(v => v.toString())).div(ONE)
             assert(
                 isClose(actual, expected),
                 `Marginal price calculation is off for iteration ${i}:\n` +
@@ -92,7 +92,7 @@ contract('MarketMaker', function(accounts) {
         // Selling of tokens is worth less than 1 Wei
         assert.equal(profit, 0)
         // User's Ether balance increased
-        assert((await etherToken.balanceOf.call(accounts[trader])).gt(buyerBalance), 'trader balance did not increase')
+        assert((await etherToken.balanceOf(accounts[trader])).gt(buyerBalance), 'trader balance did not increase')
     })
 
     it('should move price of an outcome to 1 after participants buy lots of that outcome from lmsrMarketMaker maker', async () => {
@@ -100,15 +100,13 @@ contract('MarketMaker', function(accounts) {
         const numOutcomes = 2
         const questionId = '0xf00dcafef00dcafef00dcafef00dcafef00dcafef00dcafef00dcafef00dcafe'
         const oracleAddress = accounts[5]
-        const conditionId = await getParamFromTxEvent(
-            await pmSystem.prepareCondition(oracleAddress, questionId, toBN(numOutcomes)),
-            'conditionId')
+        const conditionId = await getParamFromTxEvent(await pmSystem.prepareCondition(oracleAddress, questionId, toBN(numOutcomes)), 'conditionId')
         
-            for(const [investor, funding, tokenCountRaw] of [
-                [2, toBN(1e17), 1e18],
-                [3, toBN(1), 10],
-                [4, toBN(1), 1e18],
-            ]) {
+        for(const [investor, funding, tokenCountRaw] of [
+            [2, toBN(1e17), 1e18],
+            [3, toBN(1), 10],
+            [4, toBN(1), 1e18],
+        ]) {
             // // Calculate the address of the LMSR via nonce before it's deployed (in order to allow approve() call)
             const checksummedLMSRAddress = toChecksumAddress(keccak256(rlp.encode([lmsrMarketMakerFactory.address, nonce])).substr(26));
             await etherToken.deposit({ value: funding, from: accounts[investor] })
@@ -118,15 +116,12 @@ contract('MarketMaker', function(accounts) {
             const netOutcomeTokensSold = new Array(numOutcomes).fill(0)
             
             // Create and Fund lmsrMarketMaker
-
-            assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), funding.toString())
+            assert.equal((await etherToken.balanceOf(accounts[investor])).toString(), funding.toString())
             const feeFactor = 0  // 0%
             const lmsrMarketMaker = await getParamFromTxEvent(
-                await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, funding,
-                    { from: accounts[investor] }),
-                'lmsrMarketMaker', LMSRMarketMaker)
+                await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, funding, { from: accounts[investor] }),'lmsrMarketMaker', LMSRMarketMaker)
             nonce++
-            assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), '0')
+            assert.equal((await etherToken.balanceOf(accounts[investor])).toString(), '0')
 
             // User buys ether tokens
             const trader = 1
@@ -139,17 +134,17 @@ contract('MarketMaker', function(accounts) {
             for(const i of _.range(loopCount)) {
                 // Calculate cost of buying tokens
                 outcomeTokenAmounts = Array.from({length: numOutcomes}, (v, i) => i === outcome ? tokenCount : toBN(0))
-                cost = await lmsrMarketMaker.calcNetCost.call(outcomeTokenAmounts)
+                cost = await lmsrMarketMaker.calcNetCost(outcomeTokenAmounts)
 
                 // Buying tokens
                 await etherToken.approve(lmsrMarketMaker.address, cost, { from: accounts[trader] })
-                assert.equal(await getParamFromTxEvent(
+                assert.equal((await getParamFromTxEvent(
                     await lmsrMarketMaker.trade(outcomeTokenAmounts, cost, false, { from: accounts[trader] }), 'outcomeTokenNetCost'
-                ).then(v => v.toString()), cost.toString())
+                )).toString(), cost.toString())
 
                 netOutcomeTokensSold[outcome] += tokenCountRaw
                 const expected = lmsrMarginalPrice(funding, netOutcomeTokensSold, outcome)
-                const actual = new Decimal(await lmsrMarketMaker.calcMarginalPrice.call(toBN(outcome)).then(v => v.toString())).div(ONE)
+                const actual = new Decimal((await lmsrMarketMaker.calcMarginalPrice(toBN(outcome))).toString()).div(ONE)
                 assert(
                     isClose(actual, expected) || expected.toString() == 'NaN',
                     `Marginal price calculation is off for iteration ${i}:\n` +
@@ -179,18 +174,18 @@ contract('MarketMaker', function(accounts) {
         // // Calculate the address of the LMSR via nonce before it's deployed (in order to allow approve() call)
         const checksummedLMSRAddress = toChecksumAddress(keccak256(rlp.encode([lmsrMarketMakerFactory.address, nonce])).substr(26));
         const funding = toBN(1e18)
+        const feeFactor = toBN(0)  // 0%
         await etherToken.deposit({ value: funding, from: accounts[investor] })
         await etherToken.approve(checksummedLMSRAddress, funding, { from: accounts[investor] })
-        assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), funding.toString())
+        assert.equal(await etherToken.balanceOf(accounts[investor]).then(v => v.toString()), funding.toString())
 
-        const feeFactor = toBN(0)  // 0%
         const lmsrMarketMaker = await getParamFromTxEvent(
             await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, funding,
                 { from: accounts[investor] }),
-            'lmsrMarketMaker', LMSRMarketMaker)
-        
+            'lmsrMarketMaker', LMSRMarketMaker)        
         nonce++
-        assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), '0')
+
+        assert.equal((await etherToken.balanceOf(accounts[investor])).toString(), '0')
 
         const trader = 6
         const initialOutcomeTokenCount = toBN(1e18)
@@ -203,268 +198,26 @@ contract('MarketMaker', function(accounts) {
 
         // User trades with the lmsrMarketMaker
         const tradeValues = [5e17, -1e18, -1e17, 2e18].map(toBN)
-        const cost = await lmsrMarketMaker.calcNetCost.call(tradeValues)
+        const cost = await lmsrMarketMaker.calcNetCost(tradeValues)
         if(cost.gtn(0)) await etherToken.approve(lmsrMarketMaker.address, cost, { from: accounts[trader] })
 
         await pmSystem.setApprovalForAll(lmsrMarketMaker.address, true, { from: accounts[trader] })
 
-        assert.equal(await getParamFromTxEvent(
+        assert.equal((await getParamFromTxEvent(
             await lmsrMarketMaker.trade(tradeValues, cost, false, { from: accounts[trader] }), 'outcomeTokenNetCost'
-        ).then(v => v.toString()), cost.toString())
+        )).toString(), cost.toString())
 
         // All state transitions associated with trade have been performed
         for(const [tradeValue, i] of tradeValues.map((v, i) => [v, i])) {
-            assert.equal(await pmSystem.balanceOf.call(accounts[trader], soliditySha3(
+            assert.equal((await pmSystem.balanceOf(accounts[trader], soliditySha3(
                 { t: 'address', v: etherToken.address },
                 { t: 'bytes32', v: soliditySha3(
                     { t: 'bytes32', v: conditionId },
                     { t: 'uint', v: 1 << i },
                 )}
-            )).then(v => v.toString()), initialOutcomeTokenCount.add(tradeValue))
+            ))).toString(), initialOutcomeTokenCount.add(tradeValue))
         }
 
-        assert.equal(await etherToken.balanceOf.call(accounts[trader]).then(v => v.toString()), initialWETH9Count.sub(cost).toString())
-    })
-})
-
-
-contract('LMSRMarketMaker', function (accounts) {
-    let pmSystem
-    let etherToken
-    let conditionId
-    let lmsrMarketMakerFactory
-    let centralizedOracle
-    let questionId = 100
-    const numOutcomes = 2
-    let nonce = 0x01
-
-    before(async () => {
-        pmSystem = await PredictionMarketSystem.deployed()
-        etherToken = await WETH9.deployed()
-        lmsrMarketMakerFactory = await LMSRMarketMakerFactory.deployed()
-    })
-
-    beforeEach(async () => {
-        // create event
-        centralizedOracle = accounts[1]
-        questionId++
-        conditionId = await getParamFromTxEvent(
-            await pmSystem.prepareCondition(centralizedOracle, toHex(questionId), toBN(numOutcomes)),
-            'conditionId')
-    })
-
-    it('can be created and closed', async () => {
-        // Create lmsrMarketMaker
-        const buyer = 5
-        const funding = toBN(100)
-        const feeFactor = toBN(0)
-                
-        // // Calculate the address of the LMSR via nonce before it's deployed (in order to allow approve() call)
-        const checksummedLMSRAddress = toChecksumAddress(keccak256(rlp.encode([lmsrMarketMakerFactory.address, nonce])).substr(26));
-        await etherToken.deposit({ value: funding, from: accounts[buyer] })
-        await etherToken.approve(checksummedLMSRAddress, funding, { from: accounts[buyer] }) 
-
-        assert.equal(await etherToken.balanceOf.call(accounts[buyer]).then(v => v.toString()), funding.toString())
-
-        const lmsrMarketMaker = await getParamFromTxEvent(
-            await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, funding, { from: accounts[buyer] }),
-            'lmsrMarketMaker', LMSRMarketMaker
-        )
-        nonce++
-
-        // Close lmsrMarketMaker
-        await lmsrMarketMaker.close({ from: accounts[buyer] })
-
-        // LMSRMarketMaker can only be closed once
-        await assertRejects(lmsrMarketMaker.close({ from: accounts[buyer] }), 'lmsrMarketMaker closed twice')
-
-        // Sell all outcomes
-        await pmSystem.setApprovalForAll(lmsrMarketMaker.address, true, { from: accounts[buyer] })
-        await pmSystem.mergePositions(etherToken.address, '0x00', conditionId, [...Array(numOutcomes).keys()].map(i => toBN(1 << i)), funding, { from: accounts[buyer] })
-        assert.equal(await etherToken.balanceOf.call(accounts[buyer]).then(v => v.toString()), funding.toString())
-    })
-
-    it('should allow buying and selling', async () => {
-        // create lmsrMarketMaker
-        const investor = 0
-
-        const feeFactor = toBN(5e16)  // 5%
-        const funding = toBN(1e18)
-
-        // // Calculate the address of the LMSR via nonce before it's deployed (in order to allow approve() call)
-        const checksummedLMSRAddress = toChecksumAddress(keccak256(rlp.encode([lmsrMarketMakerFactory.address, nonce])).substr(26));
-        await etherToken.deposit({ value: funding, from: accounts[investor] })
-        await etherToken.approve(checksummedLMSRAddress, funding, { from: accounts[investor] }) 
-        assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), funding.toString())
-     
-        const lmsrMarketMaker = await getParamFromTxEvent(
-            await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, funding, { from: accounts[investor] }),
-            'lmsrMarketMaker', LMSRMarketMaker
-        )
-        nonce++
-
-        assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), '0')
-
-        // Buy outcome tokens
-        const buyer = 1
-        const outcome = 0
-        const positionId = soliditySha3(
-                { t: 'address', v: etherToken.address },
-                { t: 'bytes32', v: soliditySha3(
-                    { t: 'bytes32', v: conditionId },
-                    { t: 'uint', v: 1 << outcome },
-                )}
-            )
-        const tokenCount = toBN(1e15)
-        let outcomeTokenAmounts = Array.from({length: numOutcomes}, (v, i) => i === outcome ? tokenCount : toBN(0))
-        const outcomeTokenCost = await lmsrMarketMaker.calcNetCost.call(outcomeTokenAmounts)
-
-        let fee = await lmsrMarketMaker.calcMarketFee.call(outcomeTokenCost)
-        assert.equal(fee.toString(), outcomeTokenCost.muln(5).divn(100).toString())
-
-        const cost = fee.add(outcomeTokenCost)
-        await etherToken.deposit({ value: cost, from: accounts[buyer] })
-        assert.equal(await etherToken.balanceOf.call(accounts[buyer]).then(v => v.toString()), cost.toString())
-
-        await etherToken.approve(lmsrMarketMaker.address, cost, { from: accounts[buyer] })
-        assert.equal(await getParamFromTxEvent(
-            await lmsrMarketMaker.trade(outcomeTokenAmounts, cost, false, { from: accounts[buyer] }), 'outcomeTokenNetCost'
-        ), outcomeTokenCost.toString())
-
-        assert.equal(await pmSystem.balanceOf.call(accounts[buyer], positionId).then(v => v.toString()), tokenCount.toString())
-        assert.equal(await etherToken.balanceOf.call(accounts[buyer]).then(v => v.toString()), 0)
-
-        // Sell outcome tokens
-        outcomeTokenAmounts = Array.from({length: numOutcomes}, (v, i) => i === outcome ? tokenCount.neg() : toBN(0))
-        const outcomeTokenProfit = (await lmsrMarketMaker.calcNetCost.call(outcomeTokenAmounts)).neg()
-        fee = await lmsrMarketMaker.calcMarketFee.call(outcomeTokenProfit)
-        const profit = outcomeTokenProfit.sub(fee)
-
-        await pmSystem.setApprovalForAll(lmsrMarketMaker.address, true, { from: accounts[buyer] })
-        assert.equal(await getParamFromTxEvent(
-            await lmsrMarketMaker.trade(outcomeTokenAmounts, profit.neg(), false, { from: accounts[buyer] }), 'outcomeTokenNetCost'
-        ).then(v => v.neg().toString()), outcomeTokenProfit.toString())
-
-        assert.equal(await pmSystem.balanceOf.call(accounts[buyer], positionId).then(v => v.toString()), '0')
-        assert.equal(await etherToken.balanceOf.call(accounts[buyer]).then(v => v.toString()), profit.toString())
-    })
-
-    it('should allow short selling', async () => {
-        // create lmsrMarketMaker
-        const investor = 7
-
-        const feeFactor = toBN(50000)  // 5%
-        const funding = toBN(1e18)
-        // // Calculate the address of the LMSR via nonce before it's deployed (in order to allow approve() call)
-        const checksummedLMSRAddress = toChecksumAddress(keccak256(rlp.encode([lmsrMarketMakerFactory.address, nonce])).substr(26));
-        await etherToken.deposit({ value: funding, from: accounts[investor] })
-        await etherToken.approve(checksummedLMSRAddress, funding, { from: accounts[investor] }) 
-
-        const lmsrMarketMaker = await getParamFromTxEvent(
-            await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, funding, { from: accounts[investor] }),
-            'lmsrMarketMaker', LMSRMarketMaker
-        )
-        nonce++
-
-        assert.equal(await etherToken.balanceOf.call(accounts[investor]).then(v => v.toString()), '0')
-
-        // Short sell outcome tokens
-        const buyer = 7
-        const outcome = 0
-        const differentOutcome = 1
-        const differentPositionId = soliditySha3(
-                { t: 'address', v: etherToken.address },
-                { t: 'bytes32', v: soliditySha3(
-                    { t: 'bytes32', v: conditionId },
-                    { t: 'uint', v: 1 << differentOutcome },
-                )}
-            )
-        const tokenCount = toBN(1e15)
-        const outcomeTokenAmounts = Array.from({length: numOutcomes}, (v, i) => i !== outcome ? tokenCount : toBN(0))
-        const outcomeTokenCost = await lmsrMarketMaker.calcNetCost.call(outcomeTokenAmounts)
-        const fee = await lmsrMarketMaker.calcMarketFee.call(outcomeTokenCost)
-        const cost = outcomeTokenCost.add(fee)
-
-        await etherToken.deposit({ value: cost, from: accounts[buyer] })
-        assert.equal(await etherToken.balanceOf.call(accounts[buyer]).then(v => v.toString()), cost.toString())
-        await etherToken.approve(lmsrMarketMaker.address, cost, { from: accounts[buyer] })
-
-        assert.equal(
-            await getParamFromTxEvent(
-                await lmsrMarketMaker.trade(outcomeTokenAmounts, cost, false, { from: accounts[buyer] }),
-                'outcomeTokenNetCost'
-            ).then(v => v.toString()), outcomeTokenCost.toString())
-
-        assert.equal(await etherToken.balanceOf.call(accounts[buyer]).then(v => v.toString()), '0')
-        assert.equal(await pmSystem.balanceOf.call(accounts[buyer], differentPositionId).then(v => v.toString()), tokenCount.toString())
-    })
-
-    it('trading stress testing', async () => {
-        const MAX_VALUE = toBN(2).pow(toBN(256)).subn(1)
-
-        const trader = 9
-        const feeFactor = toBN(0)
-        const funding = toBN(1e16)
-
-        // // Calculate the address of the LMSR via nonce before it's deployed (in order to allow approve() call)
-        const checksummedLMSRAddress = toChecksumAddress(keccak256(rlp.encode([lmsrMarketMakerFactory.address, nonce])).substr(26));
-        await etherToken.deposit({ value: funding, from: accounts[trader] })
-        await etherToken.approve(checksummedLMSRAddress, funding, { from: accounts[trader] }) 
-
-        const lmsrMarketMaker = await getParamFromTxEvent(
-            await lmsrMarketMakerFactory.createLMSRMarketMaker(pmSystem.address, etherToken.address, [conditionId], feeFactor, funding, { from: accounts[trader] }),
-            'lmsrMarketMaker', LMSRMarketMaker
-        )
-        nonce++
-
-        const positionIds = [...Array(numOutcomes).keys()].map(i => soliditySha3(
-                { t: 'address', v: etherToken.address },
-                { t: 'bytes32', v: soliditySha3(
-                    { t: 'bytes32', v: conditionId },
-                    { t: 'uint', v: 1 << i },
-                )}
-            ))
-
-        // Get ready for trading
-        const tradingStipend = toBN(1e19)
-        await etherToken.deposit({ value: tradingStipend.muln(2), from: accounts[trader] })
-        await etherToken.approve(pmSystem.address, tradingStipend, { from: accounts[trader] })
-        await pmSystem.splitPosition(etherToken.address, '0x00', conditionId, [...Array(numOutcomes).keys()].map(i => 1 << i), tradingStipend, { from: accounts[trader] })
-
-        // Allow all trading
-        await etherToken.approve(lmsrMarketMaker.address, MAX_VALUE, { from: accounts[trader] })
-        await pmSystem.setApprovalForAll(lmsrMarketMaker.address, true, { from: accounts[trader] })
-
-        for(let i = 0; i < 10; i++) {
-            const outcomeTokenAmounts = randnums(-1e16, 1e16, numOutcomes).map(n => toBN(n.valueOf()))
-            const netCost = await lmsrMarketMaker.calcNetCost.call(outcomeTokenAmounts)
-
-            const lmsrMarketMakerOutcomeTokenCounts = await Promise.all(positionIds.map(positionId =>
-                pmSystem.balanceOf.call(lmsrMarketMaker.address, positionId)))
-
-            const lmsrMarketMakerCollateralTokenCount = await etherToken.balanceOf.call(lmsrMarketMaker.address)
-
-            let txResult;
-            try {
-                txResult = await lmsrMarketMaker.trade(outcomeTokenAmounts, netCost, false, { from: accounts[trader] })
-            } catch(e) {
-                throw new Error(`trade ${ i } with input ${
-                    outcomeTokenAmounts
-                } and limit ${
-                    netCost
-                } failed while lmsrMarketMaker has:\n\n${
-                    lmsrMarketMakerOutcomeTokenCounts.map(c => c.toString()).join('\n')
-                }\n\nand ${
-                    lmsrMarketMakerCollateralTokenCount.toString()
-                }: ${
-                    e.message
-                }`)
-            }
-
-            if(txResult)
-                assert.equal(
-                    (await getParamFromTxEvent(txResult, 'outcomeTokenNetCost')).toString(),
-                    netCost.toString())
-        }
+        assert.equal((await etherToken.balanceOf(accounts[trader])).toString(), initialWETH9Count.sub(cost).toString())
     })
 })
